Add balance to the connect context
