+++ packages/web3-react/src/hooks/useBalance.tsx
@@ -15,57 +15,13 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
-import { Balance } from '@alephium/web3/dist/src/api/api-alephium'
-import { useCallback, useEffect, useState } from 'react'
+import { useMemo } from 'react'
 import { useAlephiumConnectContext } from '../contexts/alephiumConnect'
-import { SubscribeOptions, Subscription, isBalanceEqual, node, subscribeToTxStatus } from '@alephium/web3'
 
 export function useBalance() {
   const context = useAlephiumConnectContext()
-  const [balance, setBalance] = useState<Balance>()
 
-  const updateBalance = useCallback(async () => {
-    const nodeProvider = context.signerProvider?.nodeProvider
-    if (nodeProvider && context.account) {
-      const newBalance = await nodeProvider.addresses.getAddressesAddressBalance(context.account.address)
-      setBalance((prevBalance) => {
-        if (prevBalance !== undefined && isBalanceEqual(prevBalance, newBalance)) {
-          return prevBalance
-        }
-        return newBalance
-      })
-    } else if (context.account === undefined) {
-      setBalance(undefined)
-    }
-  }, [context.signerProvider?.nodeProvider, context.account, setBalance])
-
-  const updateBalanceForTx = useCallback(
-    (txId: string, confirmations?: number) => {
-      const expectedConfirmations = confirmations ?? 1
-      const pollingInterval = context.account?.network === 'devnet' ? 1000 : 4000
-      const messageCallback = async (txStatus: node.TxStatus): Promise<void> => {
-        if (txStatus.type === 'Confirmed' && (txStatus as node.Confirmed).chainConfirmations >= expectedConfirmations) {
-          await updateBalance()
-        }
-      }
-      const errorCallback = (err: any, subscription: Subscription<node.TxStatus>): Promise<void> => {
-        subscription.unsubscribe()
-        console.error(`tx status subscription error: ${err}`)
-        return Promise.resolve()
-      }
-      const options: SubscribeOptions<node.TxStatus> = {
-        pollingInterval,
-        messageCallback,
-        errorCallback
-      }
-      subscribeToTxStatus(options, txId, undefined, undefined, expectedConfirmations)
-    },
-    [updateBalance, context.account?.network]
-  )
-
-  useEffect(() => {
-    updateBalance()
-  }, [updateBalance])
-
-  return { balance, updateBalanceForTx }
+  return useMemo(() => {
+    return { balance: context.balance, updateBalanceForTx: context.updateBalanceForTx }
+  }, [context.balance, context.updateBalanceForTx])
 }
